🎨 Card outsourced
